Fix the XFT build under C++, thank you Mats Lidell.
lwlib/ChangeLog addition:

2024-08-18  Aidan Kehoe  <kehoea@parhasard.net>

	* lwlib-fonts.c:
	* lwlib-fonts.c (xft_open_font_by_name):
	* lwlib-fonts.c (x_xft_text_width): Move to this file.
	* lwlib-fonts.h:
	Declare NAME for this xft_open_font_by_name as const char *, a
	better practice anyway and placates the C++ build.
	Declare x_xft_text_width here.
	* xlwmenu.c (x_xft_text_width): Remove from this file.
	* xlwtabs.c (x_xft_text_width): Remove from this file.

src/ChangeLog addition:

2024-08-18  Aidan Kehoe  <kehoea@parhasard.net>

	* fontcolor-x.c (x_print_font_instance):
	Use the appropriate format specifier for an address when printing
	an XFT font's address.
	* redisplay-x.c (x_output_string):
	Cast some members of an XRectangle structure, suppressing a
	warning from the C++ compiler.

+++ src/redisplay-x.c
@@ -1720,8 +1720,10 @@ x_output_string (struct window *w, struct display_line *dl, const Ibyte *buf,
 	    if (need_clipping)
 	      {
 		Region clip_reg = XCreateRegion();
-		XRectangle clip_box = { clip_start, ypos,
-					clip_end - clip_start, height };
+		XRectangle clip_box = { (short) clip_start, (short) ypos,
+					(unsigned short)
+                                        (clip_end - clip_start),
+                                        (unsigned short) height };
 
 		XUnionRectWithRegion (&clip_box, clip_reg, clip_reg); 
 		XftDrawSetClip (xftDraw, clip_reg);
@@ -1901,8 +1903,9 @@ x_output_string (struct window *w, struct display_line *dl, const Ibyte *buf,
 	  
 	      { /* set up clipping */
 		Region clip_reg = XCreateRegion();
-		XRectangle clip_box = { cursor_start, ypos,
-					cursor_width, height };
+		XRectangle clip_box = { (short) cursor_start, (short) ypos,
+					(unsigned short) cursor_width,
+                                        (unsigned short) height };
 	    
 		XUnionRectWithRegion (&clip_box, clip_reg, clip_reg); 
 		XftDrawSetClip (xftDraw, clip_reg);
